Creación de los métodos crearNota, menuNota, listaDeNotas, de cambio el double de crear real a float, se cambio el nombre de identificador a id, y se modifico el toString de Estudiante

+++ Nivel_1/Sistema_Notas/src/util/Herramientas.java
@@ -20,11 +20,11 @@ public class Herramientas {
 
 	}
 
-	public static double crearReal() {
+	public static float crearReal() {
 		while (true) {
 			try {
 				System.out.print(": ");
-				return entrada.nextDouble();
+				return entrada.nextFloat();
 
 			} catch (InputMismatchException e) {
 				System.out.println("Deben ingresar un numero entero\n" + e);
